Moving BatchIdToEntityHelper back and rather excluding it in services.php as it would require more refactoring
